Message contract in truffle samples

+++ samples/truffle/contracts/Number.sol
@@ -2,7 +2,7 @@
 pragma solidity >=0.5.0 <0.6.0;
 
 contract Number {
-	int private number;
+	int128 private number;
 	
     constructor() public {
 		number = 42;
